Add files via upload
Task: Upload initial version of expense tracking code

Implemented the Expense and ExpenseDatabase classes along with example usage in the main script. The code includes basic functionalities for creating, updating, and removing expenses, as well as querying expenses by ID or title. Also, a simple example scenario is provided to demonstrate how to use the code. The README has been updated with instructions on how to run the code using an IDE or the command line.